requirements.txt
discord.py
python-dotenv
PyNaCl
yt-dlp